Building a titanium app
ACS + Titanium learning, build a titanium app which contains:
two text boxes to input two numbers
four buttons for addition, subtraction, multiplication, division
operations
result label that shows the output of these operations
Another button 'save to ACS' that brings up a pop-up to give a name to
save this result and then saves this to ACS
for example, if you input 4 and 5 and clicked + then it should show 9 in
result label. Now if click on save ACS button, enter name as firstSave
and click ok to save to ACS. You should save  both operands and operator
and result in ACS
Another button 'retrieve existing results' -> when clicked on it, should
open another window and print a list of all stored results including
both operands, operators and result.
Docs:
http://docs.appcelerator.com/titanium/latest/#!/api/Titanium.Cloud
http://docs.appcelerator.com/cloud/latest/#!/guide/acs_quickstart
http://docs.appcelerator.com/cloud/latest/#!/guide/titanium